requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.6
factory-boy==3.2.1
Faker==18.13.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3
